Feature/#247

+++ algorithm/heavyhash-gate.c
@@ -70,7 +70,7 @@ static inline uint64_t xoshiro_gen(struct xoshiro_state *state) {
     return result;
 }
 
-static int compute_rank(const uint32_t A[64][64])
+static int compute_rank(const uint8_t A[64][64])
 {
     double B[64][64];
     for (int i = 0; i < 64; ++i){
@@ -104,12 +104,12 @@ static int compute_rank(const uint32_t A[64][64])
     return rank;
 }
 
-static inline bool is_full_rank(const uint32_t matrix[64][64])
+static inline bool is_full_rank(const uint8_t matrix[64][64])
 {
     return compute_rank(matrix) == 64;
 }
 
-void generate_matrix(uint32_t matrix[64][64], struct xoshiro_state *state) {
+void generate_matrix(uint8_t matrix[64][64], struct xoshiro_state *state) {
     do {
         for (int i = 0; i < 64; ++i) {
             for (int j = 0; j < 64; j += 16) {
@@ -122,14 +122,14 @@ void generate_matrix(uint32_t matrix[64][64], struct xoshiro_state *state) {
     } while (!is_full_rank(matrix));
 }
 
-void heavyhash(const uint32_t matrix[64][64], uint8_t* pdata, size_t pdata_len, uint8_t* output)
+void heavyhash(const uint8_t matrix[64][64], uint8_t* pdata, size_t pdata_len, uint8_t* output)
 {
     uint8_t hash_first[32] __attribute__((aligned(64)));
     uint8_t hash_second[32] __attribute__((aligned(64)));
     uint8_t hash_xored[32] __attribute__((aligned(64)));
 
-    uint32_t vector[64] __attribute__((aligned(64)));
-    uint32_t product[64] __attribute__((aligned(64)));
+    uint16_t vector[64] __attribute__((aligned(64)));
+    uint16_t product[64] __attribute__((aligned(64)));
 
     kt_sha3_256((uint8_t*) hash_first, 32, pdata, pdata_len);
 
@@ -139,7 +139,7 @@ void heavyhash(const uint32_t matrix[64][64], uint8_t* pdata, size_t pdata_len, 
     }
 
     for (int i = 0; i < 64; ++i) {
-        uint32_t sum = 0;
+        uint16_t sum = 0;
         for (int j = 0; j < 64; ++j) {
             sum += matrix[i][j] * vector[j];
         }
@@ -167,7 +167,7 @@ int heavyhash_test(unsigned char *pdata, const unsigned char *ptarget, uint32_t 
     
     uint32_t seed[8];
 
-    uint32_t matrix[64][64];
+    uint8_t matrix[64][64];
     struct xoshiro_state state;
 
     kt_sha3_256((uint8_t *)seed, 32, (uint8_t *)(data+1), 32);
@@ -200,7 +200,7 @@ void heavyhash_regenhash(struct work *work)
     
     uint32_t seed[8];
 
-    uint32_t matrix[64][64];
+    uint8_t matrix[64][64];
     struct xoshiro_state state;
 
     kt_sha3_256((uint8_t *)seed, 32, (uint8_t*) (data+1), 32);
@@ -228,7 +228,7 @@ bool scanhash_heavyhash(struct thr_info *thr, const unsigned char *pmidstate,
 
     const uint32_t first_nonce = ((uint32_t *)pdata)[19];
 
-    uint32_t matrix[64][64] __attribute__((aligned(64)));
+    uint8_t matrix[64][64] __attribute__((aligned(64)));
     struct xoshiro_state state;
 
     mm128_bswap32_80( edata, pdata );
